+++ react_training/src/Day3/Task3/Que3.jsx
@@ -1,20 +1,20 @@
-import React from 'react'
+import React from "react";
 
-import { Link } from 'react-router-dom';
+import { Link } from "react-router-dom";
 const Que3 = () => {
   return (
     <div>
-        <nav>
-            <ul>
-                <li>
-      <Link to='/'>Home</Link>
-      </li>
-      <li>
-      <Link to='/about'>About</Link>
-      </li>
-      </ul>
+      <nav>
+        <ul>
+          <li>
+            <Link to="/">Home</Link>
+          </li>
+          <li>
+            <Link to="/about">About</Link>
+          </li>
+        </ul>
       </nav>
     </div>
-  )
-}
-export default Que3
+  );
+};
+export default Que3;
