+++ react_training/src/Day3/Task6/Links.jsx
@@ -6,27 +6,10 @@ Add a "Go Back" button on the detail page to return to the list. */
 import React from "react";
 import { Route, BrowserRouter, Routes } from "react-router-dom";
 import Products from "./Products";
+import { products } from "./products";
 import ProductsDetails from "./ProductsDetails";
 
-function Links() {
-  const products = [
-    {
-      id: 1,
-      name: "mobile1",
-      price: 1000,
-      qty: "20",
-      details:
-        "This is the 2024 latest model that have the brand value as Iphone.",
-    },
-    {
-      id: 2,
-      name: "mobile2",
-      price: 2000,
-      qty: 100,
-      details:
-        "This is the 2023 latest model that have the brand value as Motorola",
-    },
-  ];
+const Links = () => {
   return (
     <>
       <div>Ques6</div>
