+++ react_training/src/Day3/Task6/products.js
@@ -1,0 +1,19 @@
+const products = [
+  {
+    id: 1,
+    name: "mobile1",
+    price: 1000,
+    qty: "20",
+    details:
+      "This is the 2024 latest model that have the brand value as Iphone.",
+  },
+  {
+    id: 2,
+    name: "mobile2",
+    price: 2000,
+    qty: 100,
+    details:
+      "This is the 2023 latest model that have the brand value as Motorola",
+  },
+];
+export { products };
